Hybrid: Create a connection composite to use for create and edit
